requirements.txt
# Data processing and analysis
pandas>=2.0.0                    # Data manipulation and analysis (Task 1: EDA)
numpy>=1.24.0                    # Numerical computing and array operations

# Data visualization
matplotlib>=3.7.0                # Plotting and visualization (Task 1: EDA charts)
seaborn>=0.12.0                  # Statistical data visualization (Task 1: EDA charts)

# Jupyter notebook environment
jupyter>=1.0.0                   # Jupyter notebook server
notebook>=6.5.0                  # Notebook interface
ipykernel>=6.22.0                # IPython kernel for Jupyter

# Embeddings and semantic search (Task 2)
sentence-transformers>=2.2.0     # Sentence embeddings (all-MiniLM-L6-v2 model)
faiss-cpu>=1.7.4                 # Vector similarity search (FAISS index)
scikit-learn>=1.3.0              # Machine learning utilities (normalization, etc.)

# LangChain framework
langchain>=0.1.0                 # LangChain core framework
langchain-community>=0.0.20      # LangChain community integrations
langchain-text-splitters>=0.0.1  # Text chunking utilities (Task 2: RecursiveCharacterTextSplitter)

# LLM and text generation (Task 3: RAG Generator)
transformers>=4.30.0             # HuggingFace Transformers (LLM models for generation)
torch>=2.0.0                     # PyTorch (deep learning framework for transformers)
accelerate>=0.20.0               # HuggingFace Accelerate (optimized model loading and inference)

